requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
python-multipart==0.0.7
pillow==10.3.0
inference-sdk==1.2.1
pyspellchecker==0.7.0
opencv-python==4.9.0.80
numpy==1.26.4
pytesseract==0.3.10